Moved ChangePasswordSerializer
